requirements.txt
python-dotenv
python-telegram-bot
pillow
requests